# core-lending-collateral-management-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-collateral-management
  description: Collateral & Guarantee Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/collateral-assets/{collateralAssetId}/vehicle-details/{vehicleCollateralId}:
    get:
      tags:
        - VehicleCollateral
      summary: Get Vehicle Collateral Details by ID
      operationId: getById_11
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: vehicleCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleCollateralDTO'
    put:
      tags:
        - VehicleCollateral
      summary: Update Vehicle Collateral Details
      operationId: update_11
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: vehicleCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCollateralDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleCollateralDTO'
    delete:
      tags:
        - VehicleCollateral
      summary: Delete Vehicle Collateral Details
      operationId: delete_11
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: vehicleCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/collateral-assets/{collateralAssetId}/securities-details/{securitiesCollateralId}:
    get:
      tags:
        - SecuritiesCollateral
      summary: Get Securities Collateral Details by ID
      operationId: getById_12
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: securitiesCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SecuritiesCollateralDTO'
    put:
      tags:
        - SecuritiesCollateral
      summary: Update Securities Collateral Details
      operationId: update_12
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: securitiesCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecuritiesCollateralDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SecuritiesCollateralDTO'
    delete:
      tags:
        - SecuritiesCollateral
      summary: Delete Securities Collateral Details
      operationId: delete_12
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: securitiesCollateralId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/collateral-assets/{collateralAssetId}/vehicle-details:
    get:
      tags:
        - VehicleCollateral
      summary: List or Search Vehicle Collateral Details for an Asset
      operationId: findAll_11
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestVehicleCollateralDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseVehicleCollateralDTO'
    post:
      tags:
        - VehicleCollateral
      summary: Create Vehicle Collateral Details under the given Asset
      operationId: create_11
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCollateralDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleCollateralDTO'
  /api/v1/collateral-assets/{collateralAssetId}/securities-details:
    get:
      tags:
        - SecuritiesCollateral
      summary: List or Search Securities Collateral Details for an Asset
      operationId: findAll_12
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSecuritiesCollateralDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseSecuritiesCollateralDTO'
    post:
      tags:
        - SecuritiesCollateral
      summary: Create Securities Collateral Details under the given Asset
      operationId: create_12
      parameters:
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecuritiesCollateralDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SecuritiesCollateralDTO'
...
components:
  schemas:
...
    VehicleCollateralDTO:
      required:
        - collateralAssetId
        - vehicleType
      type: object
      properties:
        vehicleCollateralId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        vehicleType:
          type: string
          enum:
            - CAR
            - TRUCK
            - MOTORCYCLE
            - BOAT
            - AIRCRAFT
            - RV
            - COMMERCIAL_VEHICLE
            - AGRICULTURAL_VEHICLE
            - CONSTRUCTION_VEHICLE
            - OTHER
        vin:
          maxLength: 17
          minLength: 0
          type: string
        make:
          maxLength: 100
          minLength: 0
          type: string
        model:
          maxLength: 100
          minLength: 0
          type: string
        year:
          maximum: 2100
          minimum: 1900
          type: integer
          format: int32
        color:
          maxLength: 50
          minLength: 0
          type: string
        engineNumber:
          maxLength: 100
          minLength: 0
          type: string
        chassisNumber:
          maxLength: 100
          minLength: 0
          type: string
        registrationNumber:
          maxLength: 50
          minLength: 0
          type: string
        registrationDate:
          type: string
          format: date
        registrationExpiryDate:
          type: string
          format: date
        mileage:
          minimum: 0
          type: integer
          format: int32
        fuelType:
          maxLength: 50
          minLength: 0
          type: string
        transmissionType:
          maxLength: 50
          minLength: 0
          type: string
        engineCapacity:
          minimum: 0
          type: integer
          format: int32
        seatingCapacity:
          minimum: 1
          type: integer
          format: int32
        condition:
          maxLength: 50
          minLength: 0
          type: string
        purchasePrice:
          minimum: 0
          exclusiveMinimum: false
          type: number
        purchaseDate:
          type: string
          format: date
        dealerName:
          maxLength: 255
          minLength: 0
          type: string
        isFinanced:
          type: boolean
        existingLoanAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        existingLender:
          maxLength: 255
          minLength: 0
          type: string
        modifications:
          maxLength: 5000
          minLength: 0
          type: string
        accidentHistory:
          maxLength: 5000
          minLength: 0
          type: string
        serviceHistory:
          maxLength: 5000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SecuritiesCollateralDTO:
      required:
        - collateralAssetId
        - securityType
      type: object
      properties:
        securitiesCollateralId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        securityType:
          type: string
          enum:
            - STOCK
            - BOND
            - MUTUAL_FUND
            - ETF
            - CERTIFICATE_OF_DEPOSIT
            - TREASURY
            - MONEY_MARKET
            - OTHER
        isin:
          maxLength: 12
          minLength: 0
          type: string
        cusip:
          maxLength: 9
          minLength: 0
          type: string
        tickerSymbol:
          maxLength: 10
          minLength: 0
          type: string
        securityName:
          maxLength: 255
          minLength: 0
          type: string
        issuerName:
          maxLength: 255
          minLength: 0
          type: string
        issuerCountry:
          maxLength: 3
          minLength: 0
          type: string
        quantity:
          minimum: 0
          exclusiveMinimum: true
          type: number
        faceValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        purchasePrice:
          minimum: 0
          exclusiveMinimum: false
          type: number
        purchaseDate:
          type: string
          format: date
        currentMarketPrice:
          minimum: 0
          exclusiveMinimum: false
          type: number
        priceDate:
          type: string
          format: date
        currency:
          maxLength: 3
          minLength: 0
          type: string
        maturityDate:
          type: string
          format: date
        couponRate:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        dividendYield:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        creditRating:
          maxLength: 10
          minLength: 0
          type: string
        exchange:
          maxLength: 100
          minLength: 0
          type: string
        custodianName:
          maxLength: 255
          minLength: 0
          type: string
        custodianAccount:
          maxLength: 100
          minLength: 0
          type: string
        isPledged:
          type: boolean
        pledgeDate:
          type: string
          format: date
        pledgeReference:
          maxLength: 100
          minLength: 0
          type: string
        haircutPercentage:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        marginCallThreshold:
          minimum: 0
          exclusiveMinimum: false
          type: number
        restrictions:
          maxLength: 5000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestVehicleCollateralDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/VehicleCollateralDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseVehicleCollateralDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/VehicleCollateralDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    FilterRequestSecuritiesCollateralDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SecuritiesCollateralDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseSecuritiesCollateralDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/SecuritiesCollateralDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...